strnstr done

+++ courses/cunix2/libft/src/ft_memccpy.c
@@ -1,12 +1,12 @@
 #include "libft.h"
 
-void *ft_memccpy(void *dest, const void *source, char ch, int count)
+void *ft_memccpy(void *dest, const void *source, char ch, size_t count)
 {
     if (dest==source)
     {
         exit(-1);//error exception
     }
-    
+
     char *tmp_dest = (char *)dest;
     char *tmp_source = (char *)source;
