+++ courses/cunix2/libft/src/ft_strlen.c
@@ -1,6 +1,6 @@
 #include "libft.h"
 
-int  ft_strlen(char *str)
+size_t  ft_strlen(char *str)
 {
     int len = 0;
     while (*str++ != '\0')
